add pages component
