+++ src/components/Pages/Course/Course.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import CourseHeader from "./CourseHeader/CourseHeader";
 import CourseDetailHolder from "./CourseDetailHolder/CourseDetailHolder";
