+++ src/components/Pages/Course/CourseDetailHolder/CourseDetail/CourseDetail.jsx
@@ -1,12 +1,13 @@
-import React from "react";
 import FilterBox from "./FilterBox/FilterBox";
 import Courses from "./Courses/Courses";
 
 const CourseDetail = () => {
-  return <div className="h-full w-3/4">
-    <FilterBox />
-    <Courses />
-  </div>;
+  return (
+    <div className="h-full w-3/4">
+      <FilterBox />
+      <Courses />
+    </div>
+  );
 };
 
 export default CourseDetail;
