+++ src/components/Pages/Course/CourseDetailHolder/CourseDetail/Courses/Courses.jsx
@@ -1,10 +1,9 @@
-import React from "react";
 import Cards from "./Cards/Cards";
 
 const Courses = () => {
   return (
     <div className="w-full h-full my-4 flexItemBetween flex-wrap gap-8">
-      <Cards /> 
+      <Cards />
     </div>
   );
 };
