+++ src/components/Pages/Landing/index.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Header from "./Header/Header";
 
 const Landing = () => {
